Resolve "Changes in point calculation"

See merge request cse1105/2021-2022/team-repositories/oopp-group-13/repository-template!136

+++ commons/src/main/java/commons/EstimateQuestion.java
@@ -39,7 +39,7 @@ public class EstimateQuestion extends Question {
    * 1 for correct answer
    * linearly between them for answers within the percentage boundary
    */
-  public float calculateHowClose(long guessedValue) {
+  public int calculateHowClose(long guessedValue) {
     return calculateHowClose(guessedValue, activity.consumption_in_wh);
   }
